physics: body: serialize type as int

+++ src/physics/body.cpp
@@ -304,7 +304,7 @@ namespace platformer2d {
 
 	bool CBody::Serialize(YAML::Emitter& Out, const EExtendableSerializer Extendable) const
 	{
-		Out << YAML::Key << "Type" << YAML::Value << std::to_underlying(BodySpec.Type);
+		Out << YAML::Key << "Type" << YAML::Value << static_cast<int>(std::to_underlying(BodySpec.Type));
 		Out << YAML::Key << "GravityScale" << YAML::Value << GravityScale;
 
 		/* Shape */
